+++ crate/spawn_play/src/system/spawn_game_object_rectify_system.rs
@@ -7,8 +7,9 @@ use derivative::Derivative;
 use derive_new::new;
 use kinematic_model::config::{Position, Velocity};
 use object_model::play::Mirrored;
+use sequence_model::loaded::SequenceId;
 use spawn_model::{
-    config::Spawn,
+    loaded::Spawn,
     play::{SpawnEvent, SpawnParent},
 };
 use team_model::play::Team;
@@ -40,6 +41,9 @@ pub struct SpawnGameObjectRectifySystemData<'s> {
     /// `Mirrored` components.
     #[derivative(Debug = "ignore")]
     pub mirroreds: WriteStorage<'s, Mirrored>,
+    /// `SequenceId` components.
+    #[derivative(Debug = "ignore")]
+    pub sequence_ids: WriteStorage<'s, SequenceId>,
     /// `Team` components.
     #[derivative(Debug = "ignore")]
     pub teams: WriteStorage<'s, Team>,
@@ -56,6 +60,7 @@ impl<'s> System<'s> for SpawnGameObjectRectifySystem {
             mut positions,
             mut velocities,
             mut mirroreds,
+            mut sequence_ids,
             mut teams,
         }: Self::SystemData,
     ) {
@@ -76,6 +81,7 @@ impl<'s> System<'s> for SpawnGameObjectRectifySystem {
             let velocity =
                 Self::velocity_rectify(&velocities, spawn, entity_parent, mirrored_parent);
             let mirrored = Self::mirrored_rectify(mirrored_parent);
+            let sequence_id = spawn.sequence_id;
 
             parent_objects
                 .insert(entity_spawned, SpawnParent::new(ev.entity_parent))
@@ -89,6 +95,9 @@ impl<'s> System<'s> for SpawnGameObjectRectifySystem {
             mirroreds
                 .insert(entity_spawned, mirrored)
                 .expect("Failed to insert `Mirrored` component.");
+            sequence_ids
+                .insert(entity_spawned, sequence_id)
+                .expect("Failed to insert `SequenceId` component.");
             if let Some(team) = team_parent {
                 teams
                     .insert(entity_spawned, team)
@@ -122,11 +131,8 @@ impl SpawnGameObjectRectifySystem {
         } else {
             spawn_position.x
         };
-        let mut position = Position::new(
-            spawn_position_x as f32,
-            spawn_position.y as f32,
-            spawn_position.z as f32,
-        );
+        let mut position = spawn_position;
+        position.x = spawn_position_x;
         if let Some(position_parent) = positions.get(entity_parent) {
             *position += **position_parent;
         }
@@ -146,11 +152,8 @@ impl SpawnGameObjectRectifySystem {
         } else {
             spawn_velocity.x
         };
-        let mut velocity = Velocity::new(
-            spawn_velocity_x as f32,
-            spawn_velocity.y as f32,
-            spawn_velocity.z as f32,
-        );
+        let mut velocity = spawn_velocity;
+        velocity.x = spawn_velocity_x;
         if let Some(velocity_parent) = velocities.get(entity_parent) {
             *velocity += **velocity_parent;
         }
@@ -165,19 +168,18 @@ impl SpawnGameObjectRectifySystem {
 
 #[cfg(test)]
 mod tests {
-    use std::str::FromStr;
-
     use amethyst::{
         ecs::{Builder, Entity, World, WorldExt},
         shrev::EventChannel,
         Error,
     };
-    use application_test_support::AutexousiousApplication;
-    use asset_model::config::AssetSlug;
+    use application_test_support::{AssetQueries, AutexousiousApplication};
+    use assets_test::ENERGY_SQUARE_SLUG;
     use kinematic_model::config::{Position, Velocity};
     use object_model::play::Mirrored;
+    use sequence_model::loaded::SequenceId;
     use spawn_model::{
-        config::Spawn,
+        loaded::Spawn,
         play::{SpawnEvent, SpawnParent},
     };
     use team_model::play::{IndependentCounter, Team};
@@ -188,7 +190,7 @@ mod tests {
     #[test]
     fn sets_position_and_velocity_relative_to_parent() -> Result<(), Error> {
         run_test(
-            |world| spawn_entity(world, false, None),
+            |world| spawn_entity(world, false, None, None),
             |world| {
                 assert_spawn_values(
                     world,
@@ -196,6 +198,7 @@ mod tests {
                     Velocity::<f32>::new(44., 55., 66.),
                     Mirrored(false),
                     None,
+                    SequenceId::new(0),
                 )
             },
         )
@@ -204,7 +207,7 @@ mod tests {
     #[test]
     fn sets_mirrored_position_and_velocity_when_parent_mirrored() -> Result<(), Error> {
         run_test(
-            |world| spawn_entity(world, true, None),
+            |world| spawn_entity(world, true, None, None),
             |world| {
                 assert_spawn_values(
                     world,
@@ -212,6 +215,24 @@ mod tests {
                     Velocity::<f32>::new(-36., 55., 66.),
                     Mirrored(true),
                     None,
+                    SequenceId::new(0),
+                )
+            },
+        )
+    }
+
+    #[test]
+    fn sets_sequence_id_of_spawned_entity_when_specified() -> Result<(), Error> {
+        run_test(
+            |world| spawn_entity(world, false, None, Some(SequenceId::new(2))),
+            |world| {
+                assert_spawn_values(
+                    world,
+                    Position::<f32>::new(11., 22., 33.),
+                    Velocity::<f32>::new(44., 55., 66.),
+                    Mirrored(false),
+                    None,
+                    SequenceId::new(2),
                 )
             },
         )
@@ -225,6 +246,7 @@ mod tests {
                     world,
                     false,
                     Some(Team::Independent(<IndependentCounter>::new(123))),
+                    None,
                 )
             },
             |world| {
@@ -234,6 +256,7 @@ mod tests {
                     Velocity::<f32>::new(44., 55., 66.),
                     Mirrored(false),
                     Some(Team::Independent(IndependentCounter::new(123))),
+                    SequenceId::new(0),
                 )
             },
         )
@@ -251,7 +274,12 @@ mod tests {
             .run_isolated()
     }
 
-    fn spawn_entity(world: &mut World, mirrored: bool, team: Option<Team>) {
+    fn spawn_entity(
+        world: &mut World,
+        mirrored: bool,
+        team: Option<Team>,
+        sequence_id: Option<SequenceId>,
+    ) {
         let position_parent = Position::<f32>::new(1., 2., 3.);
         let velocity_parent = Velocity::<f32>::new(4., 5., 6.);
         let entity_parent = {
@@ -269,14 +297,19 @@ mod tests {
         let entity_spawned = world.create_entity().build();
         world.insert(entity_spawned);
 
+        let asset_id = AssetQueries::id(world, &*ENERGY_SQUARE_SLUG);
+        let sequence_id = sequence_id.unwrap_or_else(|| SequenceId::new(0));
         let spawn = Spawn::new(
-            AssetSlug::from_str("default/fireball")
-                .expect("Expected `default/fireball` to be a valid asset slug."),
-            Position::<i32>::new(10, 20, 30),
-            Velocity::<i32>::new(40, 50, 60),
+            asset_id,
+            Position::<f32>::new(10., 20., 30.),
+            Velocity::<f32>::new(40., 50., 60.),
+            sequence_id,
         );
 
-        send_event(world, SpawnEvent::new(spawn, entity_parent, entity_spawned));
+        send_event(
+            world,
+            SpawnEvent::new(spawn, entity_parent, entity_spawned, asset_id),
+        );
     }
 
     fn send_event(world: &mut World, spawn_event: SpawnEvent) {
@@ -290,11 +323,13 @@ mod tests {
         velocity: Velocity<f32>,
         mirrored: Mirrored,
         team: Option<Team>,
+        sequence_id: SequenceId,
     ) {
         let entity_spawned = *world.read_resource::<Entity>();
         let positions = world.read_storage::<Position<f32>>();
         let velocities = world.read_storage::<Velocity<f32>>();
         let mirroreds = world.read_storage::<Mirrored>();
+        let sequence_ids = world.read_storage::<SequenceId>();
         let teams = world.read_storage::<Team>();
         let spawn_parents = world.read_storage::<SpawnParent>();
 
@@ -307,6 +342,9 @@ mod tests {
         let mirrored_actual = mirroreds
             .get(entity_spawned)
             .expect("Expected entity to have `Mirrored` component.");
+        let sequence_id_actual = sequence_ids
+            .get(entity_spawned)
+            .expect("Expected entity to have `SequenceId` component.");
         let team_actual = teams.get(entity_spawned);
         let spawn_parent_actual = spawn_parents.get(entity_spawned);
 
@@ -314,6 +352,7 @@ mod tests {
         assert_eq!(&velocity, velocity_actual);
         assert_eq!(&mirrored, mirrored_actual);
         assert_eq!(team.as_ref(), team_actual);
+        assert_eq!(&sequence_id, sequence_id_actual);
         assert!(
             spawn_parent_actual.is_some(),
             "Expected entity to have `SpawnParent` component."